refactor: initialize
