+++ src/scripts/actions/inittialize.ts
@@ -1,0 +1,8 @@
+import { hasInjected } from '../repository/extensionState'
+import { setHasInjected } from '../usecases/extensionBehavior'
+
+export function initialize() {
+  console.log('EXTENSION INITIALIZING...')
+  setHasInjected(false) // set init state
+  console.log('Init: hasInject?', hasInjected())
+}
