+++ src/scripts/actions/willMouseEnterToPlayer.ts
@@ -1,6 +1,9 @@
 import { waitElement } from '@1natsu/wait-element'
-import { isActive } from '../libs/stateMap'
+import { isActive } from '../repository/extensionState'
 
+/**
+ * @deprecated v2までビデオタイトルのマウスホバーでプレイヤーバーの開閉機能を提供していたが提供停止した
+ */
 export const willMouseEnterToPlayer = async () => {
   const movieTitleElement = await waitElement('h1.title')
