+++ src/scripts/infrastructure/stateMap.ts
@@ -1,0 +1,2 @@
+const state = new Map()
+export default state
