+++ src/scripts/usecases/extensionBehavior.ts
@@ -1,0 +1,5 @@
+import state from '../infrastructure/stateMap'
+
+export const setHasInjected = (is: boolean) => {
+  state.set('hasInjected', is)
+}
